refactor: rename folder model to store, move all reduce functions to redecer.ts file
